feat: Channel, 조회 기능 분리
