+++ src/main/java/com/sprint/mission/discodeit/core/channel/service/ChannelService.java
@@ -5,7 +5,6 @@ import com.sprint.mission.discodeit.core.channel.dto.ChannelDto;
 import com.sprint.mission.discodeit.core.channel.dto.request.ChannelUpdateRequest;
 import com.sprint.mission.discodeit.core.channel.dto.request.PrivateChannelCreateRequest;
 import com.sprint.mission.discodeit.core.channel.dto.request.PublicChannelCreateRequest;
-import java.util.List;
 import java.util.UUID;
 
 public interface ChannelService {
@@ -13,8 +12,6 @@ public interface ChannelService {
   ChannelDto create(PublicChannelCreateRequest request);
 
   ChannelDto create(PrivateChannelCreateRequest request);
-
-  List<ChannelDto> findAllByUserId(UUID userId);
 
   ChannelDto update(UUID channelId, ChannelUpdateRequest request);
